+++ CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## HEAD
+
+**Features**
+
+**Bugs**
+
+* The Custom::ECSService custom resource now waits for newly created ECS services to stabilize [#878](https://github.com/remind101/empire/pull/878)
+
+**Performance**
+
+**Security**
+
 ## 0.10.1 (2016-06-14)
 
 **Features**